Modularize SB3 benchmarks.
Add config and script for Hopper.
